Fix timestamp with timezone comparison (#8838)
Summary:
In the original comparison the timestamp with timezone value had the millisecond portion
of the value adjusted to UTC based on the timezone id of the value.
However, this is not correct because the millisecond value is already adjusted and normalized
to be in UTC and can, therefore, be directly compared.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8838

Reviewed By: xiaoxmeng

Differential Revision: D54314295

Pulled By: kagamiori

fbshipit-source-id: b3bed29371335831e8a7b856051e29b17d1c6e53