ResultOrActions (#8841)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/8841

This will be used from readers to give them the ability of not blocking on IO.

Instead, they'll return a status indicating that they need more IO and a callback to do the IO (potentially in another thread).

There could be other reasons why a reader or data source is blocked too, and we'll take advantage of this class in upper layers too.

Reviewed By: Yuhta

Differential Revision: D54095594

fbshipit-source-id: 2854294b37395add158fe69d8f9b8a1331dbac06