Reuse Serializer (#8849)
Summary:
Adds VectorSerializer::clear(). This makes serializer trees reusable. This reduces memory allocation and more importantly allows keeping a history of data characteristics. This enables adaptivity in the use of compression and reencoding. A combination of the these reduces exchange data volume to half on the average.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8849

Reviewed By: xiaoxmeng

Differential Revision: D54173866

Pulled By: oerling

fbshipit-source-id: f8203c1c0abecf8122ed28e8e0b1a477b9d7d38c