Fix flaky tests in hash join tests (#8967)
Summary:
The flaky tests in hash join test is because the recent changes that make hash build
operators to wait for probe side to finish. This caused the race condition that the task
to complete before all the drivers from the build side to close. For hash join, the probe
side pipeline is the output pipeline and a task will be terminated if the output pipeline
completes. This caused the build side stats is not reported and fails the tests. The
operator close (from driver close) aggregates the stats to task stats.

Run the tests internally for 200 iterations.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8967

Reviewed By: mbasmanova

Differential Revision: D54554567

Pulled By: xiaoxmeng

fbshipit-source-id: 4534bcdd762f2f02cb92138fc403af040a0ffdec