Add configuration 'spark.partition_id' and refactor rand(seed) function (#8934)
Summary:
Add configuration 'spark.partition_id' and refactor fucntion rand(seed) to
consume it.

Spark partition index is required for various Spark functions such as
`rand(seed)` and `spark_partition_id`. However, when Gluten constructs
Substrait plans on the Spark driver side, the Spark partition index is
unavailable. Passing it as a function parameter becomes challenging.
Therefore, we need to specify it as a configuration to be passed to Velox
functions.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8934

Reviewed By: Yuhta

Differential Revision: D54515032

Pulled By: mbasmanova

fbshipit-source-id: a8314739d44f8d6c6c794c853db07b63013a83bb