Add custom result verifier for arbitrary() in aggregation fuzzer (#8839)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/8839

Add a custom result verifier for aggregations with arbitrary() by checking that the
arbitrary() result is one of the input element in the corresponding group. This result
verifier computes array_agg() of all inputs and checks that the arbirary() result is
contained in the array_agg() result. Since the original contains() function doesn't allow null
elements in complex types, this diff adds an $internal$contains function and use this in
the reesult verifier.

Reviewed By: spershin

Differential Revision: D53983984

fbshipit-source-id: e05554fc60cc4237c15c21cdcabdefc0bdcdf939