Add support to track the input bytes read even when we hit fragment result cache
Currently, the input bytes, positions read for a query are reported
as 0 when there is a 100% fragment result cache hit. But for some
usecases its helpful to report the input data size even when its
a fragment result cache hit.
This PR adds support to track the input data size for every fragment
we cache.