[native] Do not add extra LocalExchange for native queries with join spill enabled
Unlike Java workers, native workers do not have any special requirements on the
plan share when join spill is enabled. At the same, it is confusing that query
plan changes based on whether join spilling is enabled or not.

For reference, here is a commit that added extra LocalExchange for Java workers:
https://github.com/prestodb/presto/commit/ba98c10770a9b40687aae64c3612bdb2101d5603#diff-0d1dbdeb9958c20dabf2907e354679fdc10db52c4010f0e5d7fe38cbaa4872ee