Use uintptr_t instead of uint64_t to represent address (#8778)
Summary:
In the current `MmapArena`, `uint64_t` is used to represent an
"address", which is a bit obscure. Especially when used in map, it is
difficult to distinguish whether it represents a "length" or an
"address" (pointer).

To improve readability, `uintptr_t` is used instead of `uint64_t` when
representing an "address" (pointer).

```
   std::map<uint64_t, uint64_t> freeList_;
   std::map<uint64_t, std::unordered_set<uint64_t>> freeLookup_;
```
V.S.
```
   std::map<uintptr_t, uint64_t> freeList_;
   std::map<uint64_t, std::unordered_set<uintptr_t>> freeLookup_;
```

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8778

Reviewed By: Yuhta, pedroerp

Differential Revision: D54685231

Pulled By: mbasmanova

fbshipit-source-id: e10b62e851b0d48dfca8b37b3a33674b2282703d