Fix memory leak in Spill (#9039)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9039

In FileSpillMergeStream::create we allocate a FileSpillMergeStream on the heap using new, we call nextBatch, then we convert
it to a unique_ptr.

This leaves an opportunity for the raw pointer to be leaked if nextBatch fails (e.g. due to an OOM).  Constructing the
unique_ptr directly ensures the memory will be cleaned up.

Reviewed By: xiaoxmeng

Differential Revision: D54775852

fbshipit-source-id: 759a7a5294ba686c4a2a3ef4d3e2c3d7cd739056