Speed up MemoryPool in debug mode (#9040)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9040

The debug utilities in MemoryPoolImpl are super useful, but I noticed things were taking forever to run when I had it enabled.
Looking through the code, it looks like it's because we convert the StackTrace to a string when recording an allocation.  We
can instead store the StackTrace objects, which include the stack pointers, and convert them to strings as needed (only when
there's a failure or a leak).  This speeds up run times.

Reviewed By: xiaoxmeng

Differential Revision: D54777429

fbshipit-source-id: 3cad2693be6b0f0ec72c0712b744afa0c45a7ba3