[Verifier] Support reuse of source query output table
Add support to Presto Verifier to reuse the output tables of the source query
for control and test. Add configs --control-reuse-table and --test-reuse-table
to turn the feature on and off. If reuse-table is turned on, skip the execution
of main query as well as the setup and teardown queries. Compute checksums
on the output table of main query directly.