Fix a task early termination with group execution (#9047)
Summary:
A grouped execution task might receive no more splits for all its group before the task startup.
If that happens, the existing no more split function checks if a grouped execution task has finished
execution or not by checking if the total number of finished tasks equals to the total number of tasks (Task::checkNoMoreSplitGroupsLocked). If it equals, then it will terminate the task. If the task
hasn't started, then numDriversPerSplitGroup_ is left as zero, then the task finish condition is always
met and the task is terminated as unexpected.
This PR fixes the issue by checking if numDriversPerSplitGroup_ has been set by task startup or not
under grouped execution mode. It won't check task finish if numDriversPerSplitGroup_ is not set.
Verifies with unit test.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9047

Reviewed By: spershin

Differential Revision: D54792360

Pulled By: xiaoxmeng

fbshipit-source-id: 161439aae2b5659139aff5fdf2bf30740f89bfcb