Fix hash build spill partition calculation discovered by join fuzzer (#9071)
Summary:
When hash build calculate the partition number for input, it sizes the hash vector
based on the valid size of active rows in the input but sizes the partition number
vector using the input vector size. The valid size of active rows could be smaller
than input vector size. The join fuzzer asan detects this in the new group execution
mode.
Has run join fuzzer successfully in Meta dev mode for an hour with this fix

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9071

Reviewed By: oerling

Differential Revision: D54870046

Pulled By: xiaoxmeng

fbshipit-source-id: b9feb8e1cfa5ea8bd8326a4e6d9948b443c135ff