Vector reencoding utilities (#8875)
Summary:
Adds a utility for  changing vectors to
constants if all values are the same.

Adds a VectorMap class for mapping from a value in a vector to its position in another vector. This can be used for dictionarizing vectors, combining dictionaries and generally dealing with distinct values.

Adds a VectorMap class to map from values to positions in a vector of distinct values.

Changes the signature of estimatedSerializedSize so as not to require a shared pointer to the vector. This is convenient for a VectorMap made on a preexisting vector.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8875

Reviewed By: kevinwilfong

Differential Revision: D54266770

Pulled By: oerling

fbshipit-source-id: e4f19b61d81cb11b1fa681befa068db7e48adc97