[Iceberg] Skip data size stat on fixed-width types
The optimizer utilizes the data size stats during join enumeration to
determine whether to do a partitioned or broadcast join. The stats
collection code before this change used the manifest file's data size
to estimate the total size. This value is actually mostly incorrect
and may change depending on the file format, compression, etc.

Since fixed-width type sizes are known prior, we can skip collecting
them and let the optimizer infer the data size instead. This is the same
behavior as in the hive connector.

This change will now skip any columns which are fixed-width types
when calculating the column statistics