Fix race condition in MultiThreadedTaskCursor (#9101)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9101

In MultiThreadedTaskCursor::moveNext(), if task_ has an error we call waitForTaskDriversToFinish which expects the Task to
no longer be running.

In Task's setError method there's a brief period of time after the error is set before the Task is terminated.  If the Cursor calls
waitForTaskDriversToFinish in this window it will fail because the Task is still running.

Adding a wait for the Task to complete before calling waitForTaskDriversToFinish fixes the issue.

Reviewed By: xiaoxmeng

Differential Revision: D54961848

fbshipit-source-id: 0aca844f24a610ca700f93e3f42644dc2fa7704b