Fix use-after-move in StringIdMap (#9108)
Summary:
Using a variable after moving it is bug-prone and is not a good
practice. Although the move operation on basic types is equivalent
to copying, we should still try to avoid this usage.

In the original implementation, `entry` was moved in L82((i.e. it
should no longer be used after that), but was used again in L83.
This PR fixes this problem.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9108

Reviewed By: xiaoxmeng

Differential Revision: D55017924

Pulled By: Yuhta

fbshipit-source-id: cab5941afd58192a9b21758f100cf53a83712fc4