Fix window fuzzer test (#9153)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9153

When the window frame uses the ROWS mode, frame start and frame end define the
specific number of rows before or after the current row. Therefore, input order affects the
result of window operations using the ROWS frame. This diff makes WindowFuzzer to
always provide fully ordered inputs without ties to such operations, so that the results
are deterministic and can be compared with the reference DB and across logically
equivalentplans.

Reviewed By: kevinwilfong

Differential Revision: D55082651

fbshipit-source-id: 0a1bf9d64e41baf44c5ff4a8588b959470024ced