Fix NPE caused by testingRunArbitration (#9137)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9137

testingRunArbitration is called from within an operator to force a call to shrinkPools to simulate what would happen if
arbitration was triggered by a call to growCapacity.

It does not check if the pool was aborted due to an error during arbitration which leads to all sorts of memory issues as the
Task will be terminated as part of aborting the pool causing operators to get destructed.

Adding a check to simulate what already happens in growCapacity.

Reviewed By: xiaoxmeng

Differential Revision: D55038530

fbshipit-source-id: be443619235956a43bd08158dca276126cca4789