Fix exception thrown in custom result verifier in WindowFuzzer (#9189)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9189

The window fuzzer sometimes test functions with invalid frames (e.g., with negative frame
start or end). In this situation, the execution of the query is expected to throw. With the
current code, however, if the function uses custom result verifier, the verifier initialization
throws too and cause a test failure. This diff fixes the test failures caused by exception
thrown in custom verifier by initializing it only when the main query succeed.

Reviewed By: kgpai

Differential Revision: D55160430

fbshipit-source-id: f104992e94fd0328eda97667e8bb7de38fdce7d2