Use assertQueryFails in HBO failing queries tests
The test wasn't flaky or causing problems (the query fails 100%
of the time as designed), but using assertQueryFails instead of the try/catch
is a safer pattern because if somehow the query succeeded, we wouldn't realize
the test wasn't working as intended. Also in that case, we would skip
waitProcessQueryEvents, which could manifest as flakiness in other tests
because nobody would be holding the lock on the
InMemoryHistoryBasedPlanStatisticsProvider semaphore.