Include parameters in logged test name
Include method parameters in the test name used in
LogTestDurationListener.  This fixes the listener when there are
multiple tests in the same class with the same name, but different
numbers of parameters.

Also updated the same named tests to have different names because it is
better for tests to use distinct names.

Fixes: https://github.com/prestodb/presto/issues/20807