Fix UBSAN error in Selective readers (#9206)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9206

In a bunch of our selective readers, we have logic to treat a NULL filter as AlwaysTrue. We still pass along nullptr as the filter,
but the AlwaysTrue class is a template parameter.

This nullptr gets dereferenced, and the reference is passed to ColumnFilter where we call its member functions.  These
member functions all return constant values, and are all const final, so I'm guessing normally the compiler simply replaces
the function call with the constant value it would return.  At least that's how I'm speculating we're getting away with these
shenanigans.

When building in Release mode with UBSAN enabled it detects and flags this issue.  Fixing it so UBSAN passes and general
code cleanliness.

Reviewed By: Yuhta

Differential Revision: D55213468

fbshipit-source-id: 7c29d80605524cc9a19ba155bce2b081a64f9aff