Reduce the batch size in JoinFuzzer if the keys are all boolean (#9221)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9221

In JoinFuzzer, f the keys in a join are all boolean, we generate so many hits (due to the limited possible combinations) that it
explodes the size of the output. I've seen cases where a single iteration takes >15 minutes due to this explosion.

This limits the number of iterations we execute and the amount of coverage we get in JoinFuzzer runs.

In this change, I reduce the batch size to 10% of its typical value to control this explosion (with a single boolean key and
default settings an iteration runs in seconds compared to minutes).

I also considered simply not allowing the case where all the keys are booleans, but this would leave a gap in our coverage.

Reviewed By: xiaoxmeng

Differential Revision: D55257539

fbshipit-source-id: ae71da4b0a9887dc1e126e684abe9862f180ed09