Fix nullptr dereference in SubscriptUtil (#9220)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9220

If caching is not enabled in applyMapTyped in SubscriptUtil we don't initialize the cachedLookupTablePtr so it's a nullptr.

We then dereference the pointer casting it to the appropriate type whether or not it was initialized.  We don't use the resulting
reference unless caching is enabled (in which case it wasn't a nullptr), so it doesn't cause any problems at execution time.

It does cause UBSan to throw though.  This change cleans it up so we only deference the pointer where we use it.

Reviewed By: Yuhta

Differential Revision: D55255553

fbshipit-source-id: 65896d0163c5ff1df6d1a36f44379384f9984144