Move ValueList to functions/lib/aggregates for Spark reuse (#9213)
Summary:
The array_agg function in Spark and Presto has some inconsistent behavior
when handling null values. We need to re-implement Spark's array_agg in Velox.
To be able to reuse `ValueList`, it needs to be moved to the
`functions/lib/aggregates` directory first.

The `copyValueListToArrayWriter` method has also been moved to the
`ValueList`, making it reusable.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9213

Reviewed By: Yuhta

Differential Revision: D55406723

Pulled By: pedroerp

fbshipit-source-id: 7447cfd12f18d9ca4c474579da6a55a876f97ec1