Add setup functions for velox and build dependencies handled by package managers (#8917)
Summary:
In the current setup scripts, the packages handled by the package manager are always installed
in Centos8 and Ubuntu. This is unnecessary if we want to install a specific package.
Make these installs optional by wrapping them around a function.
Split the brew packages on MacOS based on the build vs Velox requirements.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8917

Reviewed By: xiaoxmeng

Differential Revision: D55506473

Pulled By: kgpai

fbshipit-source-id: 3e4fc31080faac6006dede1808ac9d76cb4b5f60