Grow initial capacity without checking memory arbitration (#9328)
Summary:
Some slow queries spent a lot of time waiting for memory arbitration even through it only needs
a small amount of memory for execution. This is probably blocked behind some long running
memory arbitration. If the system has free capacity, then we shouldn't block this and let it proceed.

The followup is to parallelize the local memory arbitration execution, and we only expect the query
slowdown caused by the other queries when the system is busy with global memory arbitration which
indicates we are over-provisioning memory and needs operations.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9328

Reviewed By: tanjialiang, oerling

Differential Revision: D55606121

Pulled By: xiaoxmeng

fbshipit-source-id: c89237fa00c8d2538ccc4796627bc95680935a05