[Iceberg] Add predicate in layout without pushdown_filter_enabled
The `iceberg.pushdown_filter_enabled` flag added recently
changed how the iceberg connector metadata generates
table layouts. The previous fork in the logic would cause
layouts to be generated without a partition predicate constraint
even if it existed. This can lead to table stats being incorrect
and hence incorrect statistics and poor query planning when
filtering on partitioned tables or when the config is disabled.

The original logic led to the codepath which was responsible for
filling in the predicate for tables when filter pushdown was disabled
to be unreachable. This change will now allows the predicates to
show up on partition columns in the layout.