Fix memory regression of 3-arg min_by/max_by (#9307)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9307

The previous fix of 3-arg min_by/max_by, https://github.com/facebookincubator/velox/pull/8566, introduced a memory
usage regression. Before this change, every time a pair of `(compare, value)`
is popped out of the heap, the allocated memory associated with `compare`
and/or `value` is freed immediately. After this change, the allocated memory
is not freed immediately after a pop from the heap, leading to a growth of
peak memory usage. This diff fixes this memory regression by releasing the
memory associated with `(compare, value)` once this pair is popped out of
the heap. The added unit test took peak memory of 2.32MB before the fix,
and 131.5KB after the fix.

This diff fixes https://github.com/facebookincubator/velox/issues/9254.

Reviewed By: mbasmanova

Differential Revision: D55533398

fbshipit-source-id: 870bacc229bbde3ac994bc3694059225846ba1ad