Add support for collect_list Spark aggregate function (#9231)
Summary:
The semantics of Spark's `collect_list` and Presto's `array_agg` are
generally consistent, but there are inconsistencies in the handling of null
values. Spark always ignores null values in the input, whereas Presto has a
parameter that controls whether to retain them. Moreover, Presto returns null
when all inputs are null, while Spark returns an empty array.

Because of these differences, we need to re-implement the `array_agg`
function for Spark.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9231

Reviewed By: xiaoxmeng

Differential Revision: D55639676

Pulled By: mbasmanova

fbshipit-source-id: 958471779a1fa66dba27569a6c12538ad5489f46