Fix integer overflow in Entropy aggregate when argument is constant (#9332)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9332

When the argument to Entropy is constant and we have a single group, the constant value is multiplied the number of
selected rows and then this product is added to the accumulator's "sum".  It's possible this multiplication can encounter an
integer overflow.

In other cases we add values one by one to the accumulator, first casting them to double which avoids the integer overflow.
Casting the constant value to a double when multiplying achieves the same semantics.

Reviewed By: xiaoxmeng

Differential Revision: D55614963

fbshipit-source-id: a574f704a444552b22853ac122cbde87031d6ffc