Improve error for bad row cast
Selecting from a row with two columns with the same name (or same name when all
lower case) results in an ambiguous field reference.  Throw a semantic
exception instead of a GENERIC_INTERNAL_ERRROR even though this
exception gets caught after we've finished analysis.

For example
`SELECT CAST(ROW(1, 2) AS ROW(a BIGINT, A DOUBLE)).a` will now return a
semantic exception.

This also improves the test for this case to check for a specific exception
and removes a deprecated method for testing invalid functions that catches
all runtime exceptions.