Enable result validation for collect_list Spark function in the Fuzzer (#9361)
Summary:
In https://github.com/facebookincubator/velox/issues/9231, `collect_list` is added to the disable list of `duckQueryRunner`.
However, this is unnecessary because DuckDB does not have an aggregate function
named `collect_list`, hence it would not be compared against DuckDB. This
setting is redundant.

Other than this, the results verification of `collect_list` has been set to
`nullptr`, so its results are not verified. But we can use a custom array
verifier used by Presto's `array_agg` to check the results of itself.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9361

Reviewed By: xiaoxmeng

Differential Revision: D55744044

Pulled By: mbasmanova

fbshipit-source-id: a1a94c58b2a01463261775d8b6e08b65fd986d29