Fix CSE with TRY when updating sharedSubexprValues (#9364)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9364

When evaluating shared subexpressions with TRY, we only copy the values that are valid (did not encounter exceptions) into
the sharedSubexprValues.  This means that the size of sharedSubexprValues may not have values for all the selected rows (in
particular it may have a smaller length than rows).

We need to handle this when copying values from sharedSubexprValues into the result Vector, we should only copy values
that did not encounter exceptions.  TRY will handle this upstream by appending nulls for the exceptions.

I encountered this in fuzzing, the map function produces a result Vector that is only large enough to hold values that did not
encounter exceptions and this causes an error in context.moveOrCopyResult because it was trying to copy more elements than exist.

Reviewed By: Yuhta

Differential Revision: D55719677

fbshipit-source-id: bfc0cc67f86902d742fcf5fd34798f2ba1eaf342