Fix bug in initializing global aggregations with initialized flag (#9396)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9396

https://github.com/facebookincubator/velox/pull/9067 introduced an initialized flag that we include alongside the null flag
for aggregations in the rows in RowContainer.

Global aggregations do not use a RowContainer to create rows, rather a single row is constructed manually inside
GroupingSets.  When that PR was landed this logic was not properly updated, we did not allocate enough space in the row
for the additional bits which could lead to null/initialized flags overwriting the aggregate values when there were a lot of
them.

This change fixes it so that we allocate the correct amount of space for flags for global aggregations as an immediate fix.

Longer term it would be better to centralize the logic for creating rows in RowContainer (I'd like to get us back to a working
state before attempting such a refactor).

Reviewed By: Yuhta

Differential Revision: D55823098

fbshipit-source-id: 9a15f257669273c64dc0361bf735c4472e45957d