Fix flaky test in Bitmap
The testRoundTrip unit test in Bitmap was flaky because it
relied on a random set of bytes, which hit an edge case roughly
one in every 256 runs. The Bitmap class (which is a wrapper around
Java's BitSet) drops trailing zero bytes, which means that a byte
string that ended in one or more zero bytes would be serialized to
a different length than the unit test was previously expecting.

The test has since been split into two tests to clarify this behavior.
Moreover, the randomness of the test has been eliminated. This
should eliminate the flaky test failures.

See #22393.