Fix default error for approx_set and empty_approx_set (#9399)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9399

In Presto, the default standard error for `approx_set`
(and `empty_approx_set`) is
[0.01625](https://prestodb.io/docs/current/functions/hyperloglog.html#approx_set),
which is different from `approx_distinct`
([0.023](https://prestodb.io/docs/current/functions/aggregate.html#approx_distinct)).
In Velox we are using 0.023 for both, this causes problem (in addition to the
precision loss) that when `empty_approx_set` is constant folded by the
coordinator, it has a different standard error and failing the sanity check.
Fix this by aligning the default error value with Presto.

Reviewed By: mbasmanova

Differential Revision: D55814871

fbshipit-source-id: 4ecb650d02ba004fe5b1a6369d3d44464f3427ec