Remove global -Wno-return-type from CMakeLists.txt (#9387)
Summary:
This PR makes a change to remove the globally applied `-Wno-return-type`, fixes some violations. This PR originally https://github.com/facebookincubator/velox/pull/4618 disabled it. In the case of a code bug, it may become difficult to diagnose the actual cause. An example of this is

```
#include <stdio.h>

bool test1()
{
    printf("Called inner\n");
}
void test()
{
    test1();
}

int main(int argc, char* argv[])
{
    printf("Called test\n");
    test();
    printf("Finished test\n");
    return 0;
}
```
which crashes when compiled with -O2 due to bad codegen.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9387

Reviewed By: pedroerp

Differential Revision: D55933467

Pulled By: mbasmanova

fbshipit-source-id: da748fc5df19adb3d6e295973cc12e75ac43f2c1