Fix approx_distinct result verifier for window fuzzer (#9409)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9409

The custom result verification for approx_distinct in window
operations requires a different handling from aggregation. The reason
is that the error bound for approx_distinc only applies to independent
input sets, while input sets for rows in the same partition in a window
operation are correlated. Therefore, this diff changes
ApproxDistinctResultVerifier to only take the max error in each
partition when verifying the error bound for results of window
operations.

This diff fixes https://github.com/facebookincubator/velox/issues/9347.

Reviewed By: Yuhta

Differential Revision: D55898634

fbshipit-source-id: 847a70b1bfc3870bedb768f8f86711073b32e0d2