Disable signature related functions in scheduled fuzzer runs (#9420)
Summary:
For the scheduled runs it doesn't make sense to run the signature related steps, specifically generating the signature cache might clobber (through a newer date,  not by overwriting) newer stashes on main leading to false positives.

This PR also fixes issues with the generation of the baseline signature stash only happening in the scheduled jobs. The fuzzer is now  also run on merge (against the parent commit on main) and will stash the signatures afterwards for use in PRs. The signatures stash key was amended with the sha so that we can ensure that the signatures match the merge base of the PR or generate them if they are (not) yet available.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9420

Reviewed By: kagamiori, mbasmanova

Differential Revision: D55998367

Pulled By: kgpai

fbshipit-source-id: f4198f967de718c2ae1e0029091247c654297315