Store duplicate row address in vector for join probe (#9079)
Summary:
Problem
When there are a large number of rows with the same key in the build side,
the `listJoinResults` function becomes very time-consuming.

Design
`appendNextRow`
Create a next-row-vector if it doesn't exist. Append the row address to
the next-row-vector, and store the address of the next-row-vector in the
`nextOffset_` slot for all duplicate rows.

`listJoinResults`
To retrieve the addresses of all rows with the same keys, we first obtain the
address of the first row using the hash function, then, by the `nextOffset_`,
we retrieve the address of the next-row-vector. Then, we iterate through the
next-row-vector to obtain the addresses of the remaining rows. We can utilize
SIMD instructions to accelerate the next-row-vector access.

When a row needs to be erased, if value in `nextOffset_` slot is not null, then it
will be removed from corresponding next-row-vector and set it's `nextOffset_`
slot as null.

The current design is applicable to all hash modes.

Benchmark
The results indicate that this PR can accelerate the `listJoinResults` function,
with the acceleration effect becoming more pronounced as the proportion of
rows with the same key increases.

Fixes https://github.com/facebookincubator/velox/issues/9078

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9079

Reviewed By: mbasmanova

Differential Revision: D55428528

Pulled By: Yuhta

fbshipit-source-id: dfce20c1ecad3eaddc6c5e024a3b21a800d54965