Preserve case for RowType's field name and JSON content when ``CAST``
Preserve the case of double quoted field names in RowType. When casting
from ``JSON`` to ``ROW``, treat the double quoted field name as case
sensitive and unquoted field name as case insensitive in field
matching.

Add a config property `legacy_json_cast` to control and support the
legacy behavior which do not enforce the case when matching.