Fix the Window operator with overflowed empty k-rows frames (#9476)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9476

The boundaries of Window frames are int32 integers. When the frame
boundaries given in the query exceed int32 limit, the Window
operator needs to adjust the frame bounds. However, the current
code has a bug that when the frame end is below INT32_MIN, it
adjust the frame end to 0. This is wrong because the original frame is
empty, but after the adjustment, it always include row 0. This diff fixes
this bug by setting the frame bound to -1 when the frame bound
belows INT32_MIN. A subsequent call to computeValidFrames will
check whether this frame is empty and mark it properly.

This diff fixes https://github.com/facebookincubator/velox/issues/9375.

Reviewed By: Yuhta

Differential Revision: D56085211

fbshipit-source-id: 7bc330a331e12e1997b71130ab5c8267a8e75e01