Close the preloading data sources when the task terminating (#8181)
Summary:
**Problem**
Memory leaks may occur when the split preloading feature is enabled, either the
connector thread pool is busy or the task fails or is cancelled. We've observed
instances of memory leaks reported by Gluten during sporadic executions with
the split preloading feature activated.

**Proposed solution**
To resolve the issue, I propose implementing a close function in the `AsyncSource`
class. This function would be explicitly called for all preloading splits when
the task terminating. The close function would:

1. Wait for the completion of the 'make_' function if it is executing in the
thread pool.
2. Reset the 'make_' function if it has not started yet.
3. Clean up the 'item_' if 'make_' has completed, but the result 'item_' has
not been returned to the caller (necessary in cases of task cancellation).

Fixes: https://github.com/facebookincubator/velox/issues/8198

Pull Request resolved: https://github.com/facebookincubator/velox/pull/8181

Reviewed By: pedroerp

Differential Revision: D56266520

Pulled By: Yuhta

fbshipit-source-id: 391475c0731df5e77b8acb528e452fd493fd51e4