[codemod] Remove `using namespace` from github/presto-trunk/presto-native-execution/presto_cpp/main/operators/BroadcastExchangeSource.cpp +2
Summary:
Putting `using namespace` is a [_bad_ anti-pattern](https://stackoverflow.com/questions/1452721/whats-the-problem-with-using-namespace-std) and we have multiple linters and compiler errors to prevent it. But to get global coverage of `-Wheader-hygiene`, we need to draw down existing instances.

This diff removes `using namespace` from one more header files, qualifying entities in the headers. Implementation files may gain `using namespace` which is _not_ a functional change: it merely makes explicit what was previously an implicit reliance on `using namespace`. It does, however, isolate the `using namespace` to just that implementation file, which is a significant improvement over the status quo.

Differential Revision: D56272142