Fix MinioServer to prevent startup errors due to invalid ports (#9513)
Summary:
The minio server provides an API port and a console endpoint. Originally, only one free port is used
by the tests to set the API endpoint. This can cause issues (flaky test) when the selected port
overlaps with the console default port.
This change explicitly sets the ports for the API and console to avoid any issues related to these two ports.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9513

Reviewed By: xiaoxmeng

Differential Revision: D56304909

Pulled By: Yuhta

fbshipit-source-id: 5ce524aa9cbc0088780ed58c9b068026a51ccf8b