Make copies of Vectors in Values Operator if it is parallelizable (#9186)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9186

The Values Operator takes a std::vector of RowVectors and simply returns them, it's usually used as a source in test queries.

The std::vector of RowVectors comes from the Plan.  In tests where parallelizable is enabled (it's a test-only feature), there's
a risk of concurrency issues because the std::vector is shared across Drivers running on multiple threads.

This change fixes the issue by making a deep copy of the RowVectors so that each thread so a unique copy.

I introduced a function copyPreserveEncodings() to BaseVector to do the copy to ensure that the encodings of the
RowVectors' children are preserved during the copy.  This is to ensure the intentions of test authors are not violated by
silently flattening the Vectors.

This fixes several tests that are currently broken/flaky when run with TSAN, including:
* CovarianceAggregationTest/CovarianceAggregationTest.floatSomeNulls/9
* AssertQueryBuilderTest.concurrency
* ValuesTest.valuesWithRepeatAndParallelism

Reviewed By: bikramSingh91

Differential Revision: D55147065

fbshipit-source-id: 33dcad8c53a38518c782c945621c98518c93143a