Make SpillConfig callback return string_view (#9570)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9570

Previously spill config's callback returned a const reference to a
std::string. While this works if you have an std::string built on the
stack, implicit conversions like a callback returning "" is unsafe as they get
implicitly converted to a temporary std::string, thus returning a reference to
a temporary. Newer gcc flagged this issue.

Reviewed By: xiaoxmeng

Differential Revision: D56449182

fbshipit-source-id: c59ead543b42e9dea4e589413307def402504855