make numSuspensions in ThreadState atomic (#9564)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9564

numSuspensions was added in https://github.com/facebookincubator/velox/pull/9477

That PR made it so that memory arbitration does not wait for Table Scan Operators to stop.  In this case, the Table Scan
Operator may itself trigger an arbitration during an already on going arbitration.  For this reason the boolean suspended was
converted to an integer numSuspensions as multiple requests for arbitration may occur.

Naturally, the multiple requests to enter arbitration may come from different threads.  numSuspensions will only ever be
written to by a single thread (the Driver thread), but it will be read from the thread performing arbitration.  In practice this
currently isn't an issue, we only check that numSuspensions > 0 on the thread performing arbitration, since the writes (and the logic for entering and leaving suspension) happen on the Driver thread.  However, out of an abundance of caution and to
prevent future bugs, I think numSuspensions should be atomic (like many of the other variables in ThreadState).

This issue was exposed in TSAN in the join_fuzzer, where a TableScan triggered an arbitration during an ongoing arbitration,
causing a data race with this variable.

Reviewed By: xiaoxmeng

Differential Revision: D56439014

fbshipit-source-id: e36d05305e8bba50bdacccc8e280fb9498889544