Disable the driver suspension optimization for table scan (#9569)
Summary:
In Meta internal shadow testing, we found driver suspension in table scan get output could
cause high driver queue latency which slow down the query as the suspension leave does
on-thread sleep. We might need to extend it to support yield from leave if driver is pauses so
that we do off-thread sleep. Also the suspension leave needs to take care of task terminate
and throws after that, otherwise the driver might proceed to the next op which might cause
random failure if the task has been terminated as discovered by join fuzzer.
As for now, we disable this optimization and re-enable this after we have the fixes for the above
two issues later.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9569

Reviewed By: kevinwilfong, oerling

Differential Revision: D56449266

Pulled By: xiaoxmeng

fbshipit-source-id: c60c620e0a45ad1f4f878b02c02b0327ab101faf