Fix verification of histogram over interval queries in DuckDB (#9588)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9588

There's a bug in DuckDB that prevents us from verifying queries using histogram aggregates over time intervals.

The way the aggregate is written it converts the interval to a string and then tries to parse it back later.

The way we create intervals when setting up the DuckDB table, we only set microseconds and not days/months.  When
writing out to a string DuckDB converts microseconds to hours/minutes/seconds/microseconds.  The parser seems to break
trying to read that large number of hours.

By setting days and microseconds this seems to fix the parsing issue.

(This may be fixed in a later release of DuckDB but this is an easy fix on our side)

There's another bug that results in a memory leak in DuckDB when this fails that causes fuzzer to fail when run with ASAN.
This works around that issue.

Reviewed By: xiaoxmeng

Differential Revision: D56487904

fbshipit-source-id: 96eac995956283a7a556034f21c145010b79a5a5