Configure referenceDB request timeout (#9537)
Summary:
When window fuzzer is run for longer durations with Presto as reference DB, it is
observed that some queries do not reach Presto for validation and they fail in
velox with the following error message:
```
W20240418 13:15:48.194890 1896311 AggregationFuzzerBase.cpp:470] Query failed in the reference DB
```
This appears to be because the default timeout of `1000ms` for the http requests
made to Presto is lower and increasing this to a higher value, such as `5000ms`,
fixes the issue (thanks for the fix kgpai). This PR makes the request timeout used
by the `cpr::Get` request in `PrestoQueryRunner` a program argument to
`velox_window_fuzzer_test ` and `velox_aggregation_fuzzer_test`, so it can be
configured as necessary.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9537

Reviewed By: Yuhta

Differential Revision: D56371144

Pulled By: kgpai

fbshipit-source-id: 2e03df66e20bd54325954a441f9afa27ad11a3f5