Add nightly job to track build metrics (#9460)
Summary:
This PR adds a nightly job that collect build time and binary size metrics for a debug and release build and uploads the data to conbench. For example: https://velox-conbench.voltrondata.run/runs/BM-release-8665191090-1/

The adapters also add aggregate statistics, for time the `wall_time` is of course dependent on the number of cores so there will be clear differences between release and debug as debug runs on 16 cores (because the 8 cores don't have enough disk space for a debug build).

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9460

Reviewed By: Yuhta

Differential Revision: D56496198

Pulled By: kgpai

fbshipit-source-id: e6d0c3f83d10c053e90d1ad18532ff474eee25e0