Optimize copying null bits in FlatVector::copyValuesAndNulls
Summary:
Copying of null bits in FlatVector::copyValuesAndNulls showed up at 40% of the
profile in a query that's computing a dozen or so coalesce(a, b) expressions.

Optimize by copying null bits in whole words.

After the optimization, the affected query uses 2-3x less CPU time and ~half wall time.

Reviewed By: Yuhta

Differential Revision: D56571917

fbshipit-source-id: c64f3a6efda1ea14a51393d085abcd736fb1b7a5