Improve KLL Sketch perf for non-grouped queries
The current method for calculating memory usage has a hidden cost.
Within getEstimatedKllInMemorySize we call getSerializedSizeBytes.
The code for the serialized bytes size actually serializes the entire
internal state to a byte array first before returning the length. This
is expensive and should be avoided.

I am working on a PR to the upstream library to add a less-costly method
but until released, I would like to fix this as non-grouped execution
doesn't need the memory accounting for every sketch input.