Add word_stem Presto function (#9363)
Summary:
Add snowball libstemmer v2.2.0 as one of the dependencies.
And use it to implement the word_stem() as a scalar UDF.

When using the libstemmer API, each language creates an `sb_stemmer`
instance which consumes 114 bytes, including the default 10 bytes for the output stem.
It uses the `realloc` to increase the memory block for the output stem if needed.

Fixes https://github.com/facebookincubator/velox/issues/8487

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9363

Reviewed By: amitkdutta

Differential Revision: D56059511

Pulled By: pedroerp

fbshipit-source-id: b3a66956c3809e3f3dadfc8cc7b397b7116996d5