Add parquet extension to the "other" copy of duckdb 0.8.1
Summary:
For some reason a second copy of duckdb was added to the code base even though
a copy in pypi already existed. That new copy does not statically build
the parquet extension inside. In some environments (fblearner), both libraries
are being pulled in, but the the shared library loading mechanism is working on
a first-come first served bases and therefore parquet availability is dependant
on the order of the libraries and the environment. Ideally we need to eliminate
the second copy, but for now to unblock, adding parquet to the second copy as
well.

Differential Revision: D56591586

fbshipit-source-id: 87922a8cf4c3416dc29646f1d60d28a5c45d16e1