Introduce EvalCtx::captureErrorDetails flag (#9672)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9672

Introduce a flag in EvalCtx to distinguish cases where we need to capture
exception details from cases where it is sufficient to just note that an
exception occurred.

Use the new flag to optimize TRY and TRY_CAST by skipping construction of
exceptions. This optimization applies to errors reported via
EvalCtx::setStatus.

Reviewed By: bikramSingh91

Differential Revision: D56717896

fbshipit-source-id: 1322fbc7faffa3c5d8137a67afe72200083b2fc5