Enhance lazy vector consistency check (#9687)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9687

A consistency check in lazy vectors to ensure an unloaded lazy vectors
cannot be wrapped by two different top level vectors. We recently
encountered a bug where this check was triggered in HashProbe. An
input lazy vector was temporarily encapsulated before applying a
filter, and the same vector was then encapsulated again at the same
level while generating output, which triggered a failure in this
check. This change adds support for this use case to the consistency
check and ensures that HashProbe promptly destroys the temporary
vector before encapsulating it again during output generation.

Reviewed By: mbasmanova

Differential Revision: D56842170

fbshipit-source-id: 65c5f2ad670898a39439351cbed5128c48a1ea2f