Fix optimizer performance regression for large IN
This bug occurs due to large IN lists. When computing
statistics for queries with this clause, we generate an instance
of VariableStatisticEstimate for each expression in the list.

For example, a query with

SELECT ... IN (1, 2,  ... 10_000);

generates 10k VariableStatisticEstimate instances. Then, for
each instance, we sum the statistics and distinct values
to get a final result to determine the probability of a value
occurring. Creating these VariableStatisticEstimate instances
were not the root cause of the issue.

The main problem is that when the
DIsjointRangeDomainHistogram was introduced it was designed as
immutable. When folding all the instances together through
the `reduce()` call at FilterStatsCalculator#L755, it re-creates
the internal TreeRangeSet. The set of disjoint ranges is the size
of the IN list, so we were re-creating the set 10k times. The first
with range a set size of 1, then 2, ... etc up to 10k. So the
running time of this ended up being polynomial.

The following change updates the DisjointRangeDomain histogram
to use a lazily initialized RangeSet. This prevents incurring the
high cost of re-creating the range set for every  new addition to
the IN list.

Additionally, the StatisticRange class now serializes to a byte
encoded format to decrease the amount of bytes required to
serialize plans with many filters. This is mainly useful for
when a query has thousands of filters in a complex plan and the
filters are applied to the histogram.