Fix data race when resetting lazy vector's consistency check (#9701)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9701

Lazy vector's consistency check is implemented using a member variable
`containsLazyAndIsWrapped_`. A recent change ensured that it is
reset whenever the encoding layer wrapping it is destroyed which
allows it to be wrapped again with another layer. This however
showed up as a data race under TSAN because a non-lazy vector can be
wrapped by multiple encoding layers at the same level and can
therefore be potentially  cleared concurrently when they are
destroyed. This change fixes that data race.

Reviewed By: mbasmanova

Differential Revision: D56942275

fbshipit-source-id: f42221ce92e3e822de909c486d250141cac9106e