Add arg generator for decimal type (#9634)
Summary:
When creating a decimal expression in the fuzzer test, argument types could be
generated with ArgumentTypeFuzzer. But it is not able to generate argument
types which meet the required constraints given a result type. To solve this
limitation, an argument type generator is introduced, which generates all
possible decimal input types, computes result types and stores these in a map
keyed on result type. This mapping is then used to generate input types for a
given result type.
Inspired by https://github.com/facebookincubator/velox/pull/9358.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9634

Reviewed By: kgpai

Differential Revision: D56910439

Pulled By: bikramSingh91

fbshipit-source-id: 938d28fecc4d527b95571302a054a62018915805