Enhance hierarchical addition on exception context (#9695)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9695

Currently, VeloxExceptions have the ability to add additional context
from the top most level Exception context. This is useful for
debugging exceptions thrown during expression evaluation where we
want to know the current expression that threw it and the top level
one that contains it.
With this change, the "Top-level Context" concept is replaced by
"Additional Context", and a new field called "isEssential" is added
to the ExceptionContext to specify whether the context should always
be part of the exception when traversing the hierarchy. This
modification enables adding context at various levels within the
hierarchy.
Moreover, using this new method, extra details about the operator
and relevant metadata are now included in the exception to
facilitate debugging when queries fail due to runtime exceptions.

Reviewed By: mbasmanova

Differential Revision: D56913321

fbshipit-source-id: abc56abb72b0e526d8858232031279ad7a72a238