Handle ssd cache region score overflow (#9709)
Summary:
When SsdFileTracker looks for regions to evict, it calculates average
score of all regions and picks regions with no pins and whose scores no
larger than the average as candidates. It first adds up all the scores
then divided it by the number of regions to calculate the average.
However, if the scores are large, the sum can overflow to a small
integer, leading to a scenario where none of the individual region
scores is less or equal to the average. When this happens, even when
all regions are unpinned, the ssd cache file will not be able to find
any candidates, blocking all cache writes.
Therefore, overflow needs to be handled properly.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9709

Reviewed By: xiaoxmeng

Differential Revision: D56965266

Pulled By: zacw7

fbshipit-source-id: 414e67975568618a0d6ba48b947c0de73cb8103c