Register some re-usable Presto functions for Spark (#9425)
Summary:
In Gluten, both presto functions & spark functions are registered. But, there are some conflict issues. For example, when we expect to call a vector function from spark registry, a simple function with same signature from Presto registry can be called unexpectedly. So it's better to let upper framework register its own scalar function set. If some presto functions are re-usable, we need to explicitly register it for spark as this PR proposed.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9425

Reviewed By: kagamiori

Differential Revision: D57111008

Pulled By: pedroerp

fbshipit-source-id: 8abc99989d03318b3e6925e9738883c8eb3d63d3