Treat buffer as completed if no more pages
The new exchange protocol in #21926 expects the HEAD
request to indicate the buffer is completed when
the noMorePages signal has been set and the buffer is
empty.  This commit fixes the current behavior, which
is that the coordinator expects at least one GET to
transition the output buffer to completed state before
returning that the buffer has been completed.