Add overwrite flag to simple function registration API (#9158)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/9158

The `SimpleFunctionRegistry::registerFunction(aliases, constraints)`
API used to always overwrite the function registry when the function
name and signature already exists. This diff adds an overwrite flag to
this API to control the behavior of overwriting. The overwrite flag is
true by default. The
`SimpleFunctionRegistry::registerFunction(aliases, constraints, overwrite)`
API returns a bool that is true only when all aliases are successfully
registered.

Reviewed By: bikramSingh91

Differential Revision: D55041377

fbshipit-source-id: 76cea41f98de717dd8cea83df7771226d85d8b53