Move UI to a separate module
Moves the UI build to a separate module and stores the output
of the build into a maven-compatible folder structure. The new
module POM packages the UI as a resource-only jar. This module
is now defined as a dependency in presto-main so that the files
will be on the classpath for the presto server. If the UI is
excluded during the build the web UI port will return a 404.

This module can be exluded from the build via maven using
-pl '!presto-ui'.