Use mutex lock to protect cache shutdown (#9770)
Summary:
Currently, SSD cache use atomics to handle shutdown and writeInProgress counter which are used a lock to prevent concurrent writes to a SSD file. It is better to use mutex to ease the implementation and also prevent any write to SSD file if shutdown has been triggered.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9770

Reviewed By: xiaoxmeng

Differential Revision: D57220497

Pulled By: zacw7

fbshipit-source-id: caac90c041ff095678feb30df129b6c0f2fda897