Record estimation stats during query optimization
When optimizer returns a optimized plan, it will also return the estimation of stats for each node with the plan,
However, instead of returning the exact stats which are used in optimization, it's actually recalculating the stats.
This can be a problem. For example, currently CBO returns empty stats if the aggregation step is not single for an aggregation
This means that, we will not get any CBO stats for partial and final aggregation, and all other node which are downstream of the aggregation.
In this PR, it will record the stats during query optimization. For the same node, later stats will override previous ones.