Fix the issue that the splits cannot be skipped for some special case (#9707)
Summary:
In some rare case, the tables created by other engines would put the partition key columns in the data file too, and fail to write ColumnStats for such columns. In such case, the split from a partition with NULL partition keys failed to be skipped, causing wrong results. This is because HiveConnectorUtils::testFilters() assumes the partition keys are not in the data file, therefore is unable to apply the filter on the partition key. This commit fixes this issue by also checking the partition key list even when the partition columns are in the data file.

Unlike https://github.com/facebookincubator/velox/pull/9275, this PR doesn't modify the DwrfWriter to not to write ColumnStatistics, but uses a Parquet file from a null partition in a Spark generated Iceberg table.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9707

Reviewed By: Yuhta

Differential Revision: D57417377

Pulled By: kgpai

fbshipit-source-id: d1f0b4f5f16a4a6a8943703a0706341184f3d651