Let driver go off-thread waiting for memory arbitration (#9766)
Summary:
This PR adds customized memory reclaimer for query execution and implements inside the
query context:
(1) let driver go-off thread wait for query memory arbitration to complete. When a query is under
memory arbitration, it is out of memory capacity and if we allow the query driver threads continue
to run, then it is very likely that they are blocked inside the memory arbitrator which is on-thread
blocking. This might indirect affects non-spilled query or query which is not under memory arbitration
as they hold the driver thread and the non-spilled query might see high driver queue latency. Also
note that a query might contain multiple tasks and we reclaim memory from one task at a time. This
PR returns driver a future if the associated query is under memory arbitration. With Meta internal logs
based shadowing, the memory arbitration wall-time from the spilled query has been reduced by half
with this optimization. It doesn't reduce the driver queue latency but run through with more large queries
with spilling with the same or reduced the execution time.
(2) customize the query reclaim procedure to skip reclaim from the task which is non-spillable and sort
the tasks for reclaim based on the reclaimable bytes instead of task memory capacity.

The next step is to (1) parallelize memory reclamation processing at the plan node level; (2) self-triggered
spill if the operator has already spilled and do spill by itself if its memory capacity reaches to % of the
previous peak memory usage; (3) skip the memory copy in http response handling to avoid triggering the
memory arbitration in http executor in Prestissimo.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9766

Reviewed By: oerling

Differential Revision: D57197133

Pulled By: xiaoxmeng

fbshipit-source-id: d871ad7154b934b0e896be083fb6c655ae9bf5bc