Fixes: security issues for webapp jquery and bootstrap
Presto web app CVE-related package updates and its side effect fixes. The main change is to update bootstrap to 5 and jQuery to 3.7.1 to resolve the security issues.

This cherry-pick resolves all conflicts introduced along with applying the updates to new prestodb components: QueryOverview, QueryPlanView, QueryResults, QueryStageView, QueryViewer, SQLInput, StaticQueryHeader, and SessionProps. And adds a minor change to the proxy object in webpack.config to resolve an issue with yarn serve command not working.

Co-authored-by: Roney Thomas <Roney.Thomas2@ibm.com>
Co-authored-by: Alia-Alsheha <alia.alsheha@ibm.com>
Co-authored-by: Mohammad-Linjawi <mohammad.linjawi@ibm.com>
Co-authored-by: Dawi-Alotaibi <dawi.alotaibi@ibm.com>
Co-authored-by: a-alosaimi <a.alosaimi@ibm.com>
Co-authored-by: Shibil-Rahman-P2 <shibil.rahman.p2@ibm.com>