Fix local exchange source race which might cause memory leak check failure (#9869)
Summary:
We have seen memory pool leak check failure at the end of some multi-fragment test case.
By checking the log, there is no leaking task as OperatorTestBase::tearDown waits for the
pending tasks to be destroyed. We do have pending shared_ptr reference to the leaked
memory pool. By looking at the code, it could be due to the following race in how we handle
the local exchange source shutdown (note local exchange source is used by unit test for now):
T1. task sends a request;
T2. before the local exchange sets a timeout for this request, the test quit because of fault
      injection and multi-fragment test case tearDown stops the local exchange source execution
      which clears the timeout and reset the background timeout check daemon;
T3. the local exchange source tries to set timeout for the request in T1 and create a new timeout
T4. the operator test base shutdown waits for all the created tasks to be destroyed.
T5. the memory system shutdown and found there is leaked reference to the memory pool held
       by pending timeout created by T3.

This PR fixes this by adding localExchangeSource start method which is called when we start the
the multiframent unit test case in startUp method which clears 'stop_' flag to allow to create the
background daemon on the first request. Correspondingly, if 'stop_' has been set, we shouldn't allow
to create anymore request timeout and throws if 'stop_' is set. Also when we shutdown
localExchangeSource, we set 'stop_' under lock and makes sure stop_ and background timeout check
daemon and the pending timeouts states are always consistent.

Passed Meta internal test for 10'000 times: https://www.internalfb.com/intern/testinfra/testrun/10133099196500739

Pull Request resolved: https://github.com/facebookincubator/velox/pull/9869

Reviewed By: Yuhta, tanjialiang

Differential Revision: D57550932

Pulled By: xiaoxmeng

fbshipit-source-id: a6802d25e21d8e17050fce7aa363fc7a742c8d08